- Moved the creation of endpoints from the CommunicationChannel to the IChannelType objects. This way we can add TCP discovery only for the TCP channels. - Added the EndpointConnectMessage to notify a client that another endpoint wants to connect to it. - Added the test application for the communication layer - Added the ManualDiscoverySource that allows providing endpoint discovery in a manual fashion

+++ src/utils.srconly/Applications/ExceptionTypeToEventIdMap.cs
@@ -45,7 +45,10 @@ namespace Apollo.Utils.Applications
                     { typeof(PathTooLongException), 108 },
 
                     // Exception, used in case nothing else fits
-                    { typeof(Exception), int.MaxValue }
+                    //
+                    // The maximum value of an EventID is defined as UInt16.MaxValue
+                    // (i.e. 65535) here: http://msdn.microsoft.com/en-us/library/ysx54zda.aspx
+                    { typeof(Exception), ushort.MaxValue }
                 };
 
         /// <summary>
